requirements.txt
genai
pyaudio
keyboard
paho-mqtt
gTTS